python/requirements.txt
trinsic-sdk==1.9.0
deprecation